refactor api-gateway move tests
