+++ backend/api-gateway/package.json
@@ -23,10 +23,12 @@
     "@nestjs/common": "^10.0.0",
     "@nestjs/core": "^10.0.0",
     "@nestjs/platform-express": "^10.0.0",
+    "@nestjs/swagger": "^7.3.1",
     "@types/validator": "^13.11.9",
     "class-validator": "^0.14.1",
     "reflect-metadata": "^0.2.0",
     "rxjs": "^7.8.1",
+    "swagger-ui-express": "^5.0.0",
     "uuid": "^9.0.1"
   },
   "devDependencies": {
@@ -58,7 +60,8 @@
       "json",
       "ts"
     ],
-    "rootDir": "src",
+    "rootDir": "./", 
+    "modulePaths": ["<rootDir>"], 
     "testRegex": ".*\\.spec\\.ts$",
     "transform": {
       "^.+\\.(t|j)s$": "ts-jest"
