+++ backend/api-gateway/test/app.controller.spec.ts
@@ -1,6 +1,7 @@
+import { AppService } from './../src/app.service';
+import { AppController } from './../src/app.controller';
 import { Test, TestingModule } from '@nestjs/testing';
-import { AppController } from './app.controller';
-import { AppService } from './app.service';
+
 
 describe('AppController', () => {
   let appController: AppController;
